+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- FileSystem TryDeserializeAsync
 - FileSystem DecompressAsync function
 - FileSystem CompressAsync function
 - FileSystem LoadBytesAsync functions